// player/types.d.ts
export interface Style
	extends CSS.Properties<string | number>,
		CSS.PropertiesHyphen<string | number>,
		Partial<CSSTransformSpecialParam<number>> {}

export interface StyleEntry {
	[x: string]:
		| string
		| number
		| { value: number; unit?: string; ease?: string; original: string }
		| LerpStringStyle;
}

export interface LerpStringStyle {
	value: Array<number>;
	pattern: Array<string>;
	ease?: string;
	original: string;
}

/* 
ecrire les interfaces par type d'élement en entrée
*/

export interface PersoLayerDef {
	readonly type: PersoType.LAYER;
	initial: Partial<Initial>;
	actions?: Record<string, Action | boolean>;
	emit?: Emit;
}
export interface PersoTextDef {
	readonly type: PersoType.TEXT;
	initial: Partial<Initial> & { content: string };
	actions?: Record<string, Action | boolean>;
	emit?: Emit;
}

export interface PersoImgDef {
	readonly type: PersoType.IMG | PersoType.SPRITE;
	initial: Partial<Initial> & { content: Img };
	actions?: Record<string, ImgAction>;
	media?: any;
	emit?: Emit;
}

export interface PersoSoundDef {
	readonly type: PersoType.SOUND;
	initial: { src: string };
	actions?: Record<string, Partial<Action>>;
	media?: My;
	emit?: Emit;
}

export interface PersoVideoDef {
	readonly type: PersoType.VIDEO;
	initial: Partial<Initial> & { src: string };
	actions?: Record<string, Partial<Action>>;
	media?: any;
	emit?: Emit;
}

export type PersoMediaDef = PersoVideoDef | PersoSoundDef;

export interface SoundAction {
	action: typeof START | typeof STOP;
	volume: number;
	playbackRate: number;
	transition: Transition;
}

export type Broadcast =
	| BroadcastAction
	| {
			type?: BroadcastAction;
			transition?: Transition;
			volume?: number;
			playbackRate?: number;
	  };
type BroadcastAction = typeof START | typeof PAUSE | typeof STOP;

export interface ImgAction extends Omit<Action, 'content'> {
	content?: Img;
}
export type PersoDef = PersoTextDef | PersoLayerDef | PersoImgDef | PersoVideoDef;

export interface Action {
	className?: string | ActionClassList;
	style?: Style;
	content?: Content;
	transition?: Transition;
	move?: true | string | moveAction;
	strap?: StrapType;
	broadcast?: Broadcast;
}
// ...
export interface ActionClassList {
	add?: string | string[];
	remove?: string | string[];
	toggle?: string | string[];
}

export type DeltaFn = (delta: number) => void;

export type AddToTick = (fn: DeltaFn) => () => void;

export interface Income {
	time: number;
	delta: number;
	name: string;
	seek: boolean;
	data?: any;
}

export interface TimeOptions {
	delta: number;
	options: { time?: number };
}
export type TimerCallback = (data: TimeOptions) => void;

export type PersoId = string;
// ...
export type Emit = {
	readonly [key in keyof GlobalEventHandlersEventMap]?: Eventime | Array<Eventime>;
};

export interface Eventime {
	name?: string;
	// startAt: number;
	data?: any;
	duration?: number;
	events?: Eventime[];
}

export interface Perso {
	id: string;
	type: PersoType;
	initial: Partial<Initial>;
	actions: PersoAction;
	emit?: Emit;
}

interface PersoBase extends Perso {
	parent: string;
	node: HTMLElement | SVGElement;
	style: Style;
	transform: TransformProperty;
	matrix: Matrix2D | [];
	content: Set<PersoId>;
}

export interface PersoText extends PersoBase {
	type: PersoType.TEXT;
	child: Txt;
}

export interface PersoLayer extends PersoBase {
	type: PersoType.LAYER;
	child: Layer;
}

export interface PersoImg extends PersoBase {
	type: PersoType.IMG;
	initial: Partial<Initial> & Img;
	update: (update: Partial<Initial> & Img) => void;
}

export interface PersoSprite extends PersoBase {
	type: PersoType.SPRITE;
	initial: Partial<Initial> & Img;
	update: (update: Partial<Initial> & Img) => void;
}

export interface PersoVideo extends PersoBase {
	type: PersoType.VIDEO;
	initial: Partial<Initial> & Vid;
	media: HTMLVideoElement | HTMLAudioElement;
	update: (update: Partial<Broadcast>) => void;
}

export interface PersoSound extends PersoBase {
	type: PersoType.SOUND;
	initial: Partial<Initial> & Vid;
	media: My;
	update: (update: Partial<Broadcast>) => void;
}

export interface SoundNode extends PersoSoundDef {
	media?: My;
}

export interface My extends MediaElementAudioSourceNode {
	my?: {
		connect: () => void;
		disconnect: () => void;
	};
}

export type PersoNode = PersoText | PersoLayer | PersoSprite | PersoImg | PersoSound | PersoVideo;

export interface Initial {
	tag: string;
	id: string;
	attr: any;
	style: Style;
	classStyle: Style;
	className?: string | ActionClassList;
	move: string;
}

export interface Vid {
	src: string;
}

// A revoir passer les props en style sauf src et img
export interface Img {
	img?: HTMLImageElement;
	src: string;
	fit?: string;
	ratio?: number;
	width?: number | string;
	height?: number | string;
	px?: string;
	py?: string;
}
export type ImagesCollection = Map<string, Img>;

export enum PersoType {
	TEXT = 'text',
	IMG = 'img',
	LIST = 'list',
	BLOC = 'bloc',
	ROOT = 'root',
	VIDEO = 'video',
	PROTO = 'proto',
	LAYER = 'layer',
	SPRITE = 'sprite',
	BUTTON = 'button',
	POLYGON = 'polygon',
	SOUND = 'sound',
	AUDIO = 'audio',
}